backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
geopy==2.4.1
python-dateutil==2.8.2
requests==2.31.0
